worked on the footer, it looks pretty but having a hard time getting the links to work - probably something trivial.
Created the category class and product class in the models file and set up a Django admin account to create the products that are to be displayed on the website. As of right now, I'm not able to get all of the products to display on the home page but I can see the products on their individual category page.