Created new classes, order and orderitem, to begin handling user orders.
Added category and product classes to the admin page to enhance the view and editing of stock

Also changed the API keys to work with a new pinkshift stripe account. The old keys were associated with another webstore the account is also linked to (Limited Run).

Attempted to retrieve the csrf token in a query dictionary - success